% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aer_in_geom_xml.R
\name{aer_in_geom_xml}
\alias{aer_in_geom_xml}
\title{Get the coordinate of the first cell center}
\usage{
aer_in_geom_xml(sim, path = "./MECHA/Projects/GRANAR/in/Maize_Geometry.xml")
}
\arguments{
\item{sim}{The sim output list}

\item{path}{The geometry file for MECHA}
}
\description{
Get the coordinate of the first cell center
}
\keyword{mecha}
\keyword{root}
