% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aerenchyma.R
\name{aerenchyma}
\alias{aerenchyma}
\title{Create aerenchyma}
\usage{
aerenchyma(params, rs1)
}
\arguments{
\item{params}{input parameter}

\item{rs1}{node dataframe}
}
\description{
Create aerenchyma
}
\examples{
# rs <- aerenchyma(params, rs1)

}
\keyword{root}
