% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_layer.R
\name{cell_layer}
\alias{cell_layer}
\title{Place one line of cell center

Create a list}
\usage{
cell_layer(params)
}
\arguments{
\item{params}{The input dataframe}
}
\description{
Place one line of cell center

Create a list
}
\examples{
# data_list <- cell_layer(params)

}
\keyword{layer}
\keyword{root}
