% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_voro.R
\name{cell_voro}
\alias{cell_voro}
\title{Get the cooridnates for every nodes in the voronoi}
\usage{
cell_voro(all_cells, vtess, center)
}
\arguments{
\item{all_cells}{The cellular dataframe}

\item{vtess}{the voronoi data}

\item{center}{The cross-section center}
}
\description{
Get the cooridnates for every nodes in the voronoi
}
\keyword{root}
