% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_nodes.R
\name{clear_nodes}
\alias{clear_nodes}
\title{Create node dataframe of null cells}
\usage{
clear_nodes(rs1)
}
\arguments{
\item{rs1}{node dataframe}
}
\description{
Create node dataframe of null cells
}
\keyword{root}
