% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concavety.R
\name{concavety}
\alias{concavety}
\title{Merge multiple polygon cells that are adjacent to each other and can form concave polygons.}
\usage{
concavety(data)
}
\arguments{
\item{data}{the dataframe containing the cells to merge}
}
\description{
Merge multiple polygon cells that are adjacent to each other and can form concave polygons.
}
\examples{
# new_cell <- concavety(data = cells)

}
\keyword{polygons}
\keyword{root}
