% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_anatomy.R
\name{create_anatomy}
\alias{create_anatomy}
\title{Create root anatomy}
\usage{
create_anatomy(random, num_cortex, diam_cortex, size_stele, diam_stele,
  proportion_aerenchyma, n_aerenchyma_files, n_xylem_files, diam_xylem)
}
\arguments{
\item{random}{Degree of randomness in the cell positions}
}
\description{
This function creates a 2D root cross section anatomy based on global parameters
}
\examples{
create_anatomy()

}
\keyword{root}
