% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cells.R
\name{create_cells}
\alias{create_cells}
\title{Create cells space

Place initial position of the cell center}
\usage{
create_cells(all_layers, random_fact)
}
\arguments{
\item{all_layers}{The layer dataframe}

\item{random_fact}{The random factor}
}
\description{
Create cells space

Place initial position of the cell center
}
\examples{
# all_cells = create_cells(all_layers, random_fact)

}
\keyword{root}
