% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzze_inter.R
\name{fuzze_inter}
\alias{fuzze_inter}
\title{Merge connected inter-cellular space together}
\usage{
fuzze_inter(rs1)
}
\arguments{
\item{rs1}{node data frame}
}
\description{
Merge connected inter-cellular space together
}
\keyword{root}
