% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_root_section.R
\name{get_root_section}
\alias{get_root_section}
\title{Read GRANAR xml file}
\usage{
get_root_section(path)
}
\arguments{
\item{path}{The path to the GRANAR XML files}
}
\description{
This function load the GRANAR xml file.
}
\examples{
root <- list()
root$nodes <- get_root_section("PATH_TO_XLM_FILE")
plot_anatomy(root, col = "type")
}
\keyword{root}
