% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_info.R
\name{layer_info}
\alias{layer_info}
\title{Get the coordinate of the first cell center}
\usage{
layer_info(all_layers)
}
\arguments{
\item{all_layers}{The layer dataframe}
}
\description{
Get the coordinate of the first cell center
}
\keyword{layer}
\keyword{root}
