% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_it3d.R
\name{make_it3D}
\alias{make_it3D}
\title{Make a 3D root of a few cells}
\usage{
make_it3D(sim, n_cell)
}
\arguments{
\item{sim}{The sim output list}

\item{n_cell}{The number of cell layer in the root axis}
}
\description{
Make a 3D root of a few cells
}
\keyword{root}
