% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pith.R
\name{make_pith}
\alias{make_pith}
\title{Make pith inside the stele

remove stele cells and place pith cells}
\usage{
make_pith(all_cells, params, center)
}
\arguments{
\item{all_cells}{The cellular dataframe}

\item{params}{The input dataframe}

\item{center}{The cross section center location}
}
\description{
Make pith inside the stele

remove stele cells and place pith cells
}
\examples{
# all_cells <- make_pith(all_cells, params, center)

}
\keyword{root}
