% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_as_RootScan.R
\name{output_as_RootScan}
\alias{output_as_RootScan}
\title{This function turn the GRANAR output into RootScan data.}
\usage{
output_as_RootScan(sim = NULL)
}
\arguments{
\item{sim}{The output list() generated with create_anatomy() function. Default = NULL}
}
\description{
This function turn the GRANAR output into RootScan data.
}
\examples{

data <- output_as_RootScan(sim = create_anatomy(parameters = params))
}
\keyword{RootScan}
\keyword{anatomy}
\keyword{root}
