% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack_xylem.R
\name{pack_xylem}
\alias{pack_xylem}
\title{Do circle packing to add Xylem element in stele}
\usage{
pack_xylem(all_cells, params, center)
}
\arguments{
\item{all_cells}{dataframe with the id_group of the cells}

\item{params}{input parameter}

\item{center}{center of the root cross section}
}
\description{
Do circle packing to add Xylem element in stele
}
\examples{
# rs <- septa(rs1)


}
\keyword{root}
