% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_anatomy.R
\name{plot_anatomy}
\alias{plot_anatomy}
\title{Plot root anatomy}
\usage{
plot_anatomy(sim = NULL, col = "type", leg = T)
}
\arguments{
\item{sim}{the simulation objkect, returned by 'create_anatomy.R'}

\item{col}{Parameter to choose for the coloring of the cells. Accepted arguments are 'type', 'area', 'dist', 'id_cell' and 'angle'. Default = 'type'}

\item{leg}{Display the legend; Default= TRUE}
}
\description{
This function plot the results of a 2D root cross section anatomy simulation
}
\examples{
create_anatomy("PATH_TO_XLM_FILE")

}
\keyword{root}
