% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_3d_root.R
\name{plot_granar3d}
\alias{plot_granar3d}
\title{Make a 3D plot (rgl) root of a few cells}
\usage{
plot_granar3d(root, col = "type")
}
\arguments{
\item{root}{The 3d root data}

\item{col}{what should be highlited}
}
\description{
Make a 3D plot (rgl) root of a few cells
}
\keyword{3D}
\keyword{root}
