% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_geo.R
\name{prep_geo}
\alias{prep_geo}
\title{Pre-proc for .geo file generation

Add cell wall thickness
smooth cell corners
geo file can be use in GMSH (Require GMSH https://gmsh.info/)}
\usage{
prep_geo(cross_section, cell_wall_thickness, corner_smoothing)
}
\arguments{
\item{cross_section}{The nodes dataframe of the cross section}

\item{cell_wall_thickness}{the inner wall thickness data frame per cell type}

\item{corner_smoothing}{smoothing algorithm for the cells, can be cell type specific}
}
\description{
Pre-proc for .geo file generation

Add cell wall thickness
smooth cell corners
geo file can be use in GMSH (Require GMSH https://gmsh.info/)
}
\keyword{GMSH}
\keyword{Geo}
\keyword{cell}
\keyword{wall}
