% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pv_ready.R
\name{pv_ready}
\alias{pv_ready}
\title{Format data for paraview visualisation}
\usage{
pv_ready(data)
}
\arguments{
\item{data}{node dataframe}
}
\description{
Format data for paraview visualisation
}
\examples{
# rs <- pv_ready(rs1)


}
\keyword{root}
