% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_param_xml.R
\name{read_param_xml}
\alias{read_param_xml}
\title{Read the parameters for GRANAR}
\usage{
read_param_xml(path = NULL)
}
\arguments{
\item{path}{The path to the XML files ith the parameters}
}
\description{
Read the parameters for GRANAR from an XML file
}
\examples{
# read_param_xml()

}
\keyword{granar}
