% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rondy_cortex.R
\name{rondy_cortex}
\alias{rondy_cortex}
\title{Add inter cellular space

Re-do the cortex layer and make it more round shaped}
\usage{
rondy_cortex(params, all_cells, center)
}
\arguments{
\item{params}{The input dataframe}

\item{all_cells}{The cellular dataframe}

\item{center}{The cross-section center}
}
\description{
Add inter cellular space

Re-do the cortex layer and make it more round shaped
}
\examples{
# all_cells <- rondy_cortex(params, all_cells, center)

}
\keyword{root}
