% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root_hair.R
\name{root_hair}
\alias{root_hair}
\title{Create root hair from epidermis cells

extend some cells (selected randomly) to make root hair}
\usage{
root_hair(rs1, params, center)
}
\arguments{
\item{rs1}{The nodes dataframe}

\item{params}{The input dataframe}

\item{center}{The cross-section center}
}
\description{
Create root hair from epidermis cells

extend some cells (selected randomly) to make root hair
}
\keyword{root}
