% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/septa.R
\name{septa}
\alias{septa}
\title{Make Septa in between aerenchyma lacuna}
\usage{
septa(rs1)
}
\arguments{
\item{rs1}{node dataframe}
}
\description{
Make Septa in between aerenchyma lacuna
}
\examples{
# rs <- septa(rs1)


}
\keyword{root}
