% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_anatomy_xml.R
\name{substr1<-}
\alias{substr1<-}
\title{remove one str}
\usage{
substr1(x, y) <- value
}
\arguments{
\item{x}{string}

\item{y}{pattern}

\item{value}{something}
}
\description{
remove one str
}
\keyword{string}
