% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_area.R
\name{sum_area}
\alias{sum_area}
\title{Compute the sum of the cell areas}
\usage{
sum_area(cells)
}
\arguments{
\item{cells}{the dataframe containing the cells to sum up}
}
\description{
Compute the sum of the cell areas
}
\keyword{root}
