% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_geo.R
\name{unique_point_id}
\alias{unique_point_id}
\title{remove duplicated points}
\usage{
unique_point_id(tmp)
}
\arguments{
\item{tmp}{data to filter}
}
\description{
remove duplicated points
}
\keyword{GMSH}
\keyword{Geo}
