% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vascular.R
\name{vascular}
\alias{vascular}
\title{Add the vascular element in the stele

remove stele cells and place vascular elements
add round shaped boundaries for xylem elements}
\usage{
vascular(all_cells, params, layers, center)
}
\arguments{
\item{all_cells}{The cellular dataframe}

\item{params}{The input dataframe}

\item{layers}{the layer dataframe}

\item{center}{The cross-section center}
}
\description{
Add the vascular element in the stele

remove stele cells and place vascular elements
add round shaped boundaries for xylem elements
}
\examples{
# all_cells <- vascular(all_cells, params, layers, center)

}
\keyword{root}
