% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertex.R
\name{vertex}
\alias{vertex}
\title{Prepare cell wall data

tydy cells as vertex data}
\usage{
vertex(rs1)
}
\arguments{
\item{rs1}{The nodes dataframe}
}
\description{
Prepare cell wall data

tydy cells as vertex data
}
\keyword{root}
