% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_anatomy_xml_pv.R
\name{write_anatomy_xml}
\alias{write_anatomy_xml}
\title{Write the root anatomy as an XML file.

The structure of the XL file matches the one of CellSet}
\usage{
write_anatomy_xml(sim = NULL, path = NULL)
}
\arguments{
\item{sim}{The simulation results}

\item{path}{The path where to save the xml file}
}
\description{
Write the root anatomy as an XML file.

The structure of the XL file matches the one of CellSet
}
\examples{
write_anatomy_xml(sim, path = "current_root.xml" )

}
\keyword{root}
