% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_anatomy_obj.R
\name{write_CT_OBJ}
\alias{write_CT_OBJ}
\title{write a .obj from the generated root cross section}
\usage{
write_CT_OBJ(sim, path = "cross_section.obj", membrane = T)
}
\arguments{
\item{sim}{List generated from create_anatomy()}

\item{path}{Where to write the generated export}

\item{membrane}{create cell membrane if membrane = T, create cell wall if membrane = F}
}
\description{
write a .obj from the generated root cross section
}
\examples{
# write_CT_OBJ(sim, path = "cross_section.obj)

}
\keyword{anatomy}
\keyword{root}
