% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_geo.R
\name{write_geo}
\alias{write_geo}
\title{Write .geo file for GMSH

Write .geo file from cross section
geo file can be use in GMSH (Require GMSH https://gmsh.info/)}
\usage{
write_geo(data, dim = 2, path_geo, Celldomain = F)
}
\arguments{
\item{data}{The nodes dataframe of the cross section}

\item{dim}{The dimension of the input cross section, by default = 2}

\item{path_geo}{The output file with .geo extension}

\item{Celldomain}{agrument to have a cell wall domain with filled cells (TRUE), or with hollow cells (FALSE)}
}
\description{
Write .geo file for GMSH

Write .geo file from cross section
geo file can be use in GMSH (Require GMSH https://gmsh.info/)
}
\keyword{GMSH}
\keyword{Geo}
