% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_param_xml.R
\name{write_param_xml}
\alias{write_param_xml}
\title{Write a input file 'xml' for GRANAR 

This function creates a input file 'xml' for GRANAR from dataframe}
\usage{
write_param_xml(path, params)
}
\arguments{
\item{path}{Path to the future XML file containing the different parameters for GRANAR.}

\item{params}{Table with the different parameters.}
}
\description{
Write a input file 'xml' for GRANAR 

This function creates a input file 'xml' for GRANAR from dataframe
}
\examples{

root_cross_section <- create_anatomy(parameters = param)
write_param_xml(path = "Root_id.xml", params = param)
}
\author{
Adrien Heymans and Guillaume Lobet
}
