% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_anatomy.R
\name{create_anatomy}
\alias{create_anatomy}
\title{Create root anatomy}
\usage{
create_anatomy(path = NULL, parameters = NULL, verbatim = F)
}
\arguments{
\item{path}{Path to the XML file containing the different parameters for the simulation. Not needed if 'parameter' is set.  Default = NULL.}

\item{parameters}{Table with the different parameters. Not needed if 'path' is set.  Default = NULL.}

\item{verbatim}{Display textual information aboutt he simulation. Default = NULL.}
}
\description{
This function creates a 2D root cross section anatomy based on global parameters. It is the core of GRANAR.
}
\examples{
create_anatomy("PATH_TO_XLM_FILE")

}
\keyword{anatomy}
\keyword{root}
