% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_anatomy.R
\name{create_anatomy}
\alias{create_anatomy}
\title{Generate a root cross-section

Functions to generate root cross section anatomy, based on global parameters, such as the mean size of cells and number of cell layers.}
\usage{
create_anatomy(
  path = NULL,
  parameters = NULL,
  verbatim = F,
  maturity_x = F,
  paraview = T
)
}
\arguments{
\item{path}{The path to the input file}

\item{parameters}{the input parameter table. Can be obtain by read_param_xml()}

\item{verbatim}{TRUE = Generate text to follow the simulation process/ FALSE = no text}

\item{maturity_x}{TRUE = meta-xylem are labeled as stele cell /FALSE = no change in cell labeling (default)}

\item{paraview}{TRUE = cell wall data is set to make 3D object/ FALSE = cell wall data is not compatible for 3D object}
}
\description{
Generate a root cross-section

Functions to generate root cross section anatomy, based on global parameters, such as the mean size of cells and number of cell layers.
}
\examples{
# Load input
params <- read_param_xml(path = system.file("extdata", "root_monocot.xml", package = "granar"))
# Generate anatomy
root = create_anatomy(parameters = params)
# Visualize the simulation output
plot_anatomy(root)
# Write the simulation output
write_anatomy_xml(sim = root, path = system.file("extdata", "current_root.xml", package = "granar"))

}
\keyword{root}
