% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_anatomy.R
\name{create_anatomy}
\alias{create_anatomy}
\title{GRANAR: Generate root anatomy in R

This function creates a 2D root cross section anatomy based on anatomical features translated into the parameter of the model. It is the core of GRANAR.}
\usage{
create_anatomy(path = NULL, parameters = NULL, verbatim = F,
  maturity_x = F)
}
\arguments{
\item{path}{Path to the XML file containing the different parameters for the simulation. Not needed if 'parameter' is set.  Default = NULL.}

\item{parameters}{Table with the different parameters. Not needed if 'path' is set.  Default = NULL.}

\item{verbatim}{Display textual information aboutt he simulation. Default = NULL.}

\item{maturity_x}{if True Metaxylem vessels are tagged as stele cells. Default = FALSE.}
}
\description{
GRANAR: Generate root anatomy in R

This function creates a 2D root cross section anatomy based on anatomical features translated into the parameter of the model. It is the core of GRANAR.
}
\examples{

root_cross_section <- create_anatomy(path = "PATH_TO_XLM_FILE")

# OR

params <- read_param_xml(path = "PATH_TO_XLM_FILE")
root_cross_section <- create_anatomy(parameters = params)

# To visualize the result

plot_anatomy(root_cross_section)
}
\author{
Adrien Heymans and Guillaume Lobet
}
\keyword{anatomy}
\keyword{root}
